requirements.txt
dash==2.18.2
google.ai.generativelanguage==0.6.15 
pandas==2.2.3
numpy==2.2.2
google-generativeai==0.8.4
python-dotenv==1.0.1